research_assistant_fastapi/requirements.txt
fastapi
uvicorn
sqlalchemy
faiss-cpu
pymupdf
python-dotenv
requests
groq
google-genai
python-multipart